+++ src/main/kotlin/dev/aaronhowser/mods/irregular_implements/block/plate/CollectionPlateBlock.kt
@@ -2,6 +2,8 @@ package dev.aaronhowser.mods.irregular_implements.block.plate
 
 import net.minecraft.core.BlockPos
 import net.minecraft.core.Direction
+import net.minecraft.sounds.SoundEvents
+import net.minecraft.sounds.SoundSource
 import net.minecraft.world.entity.Entity
 import net.minecraft.world.entity.item.ItemEntity
 import net.minecraft.world.level.Level
@@ -11,9 +13,11 @@ import net.neoforged.neoforge.capabilities.Capabilities
 
 class CollectionPlateBlock : BasePlateBlock() {
 
-	//TODO: Maybe make a plop sound on inserting items?
 	override fun entityInside(state: BlockState, level: Level, pos: BlockPos, entity: Entity) {
 		if (entity !is ItemEntity) return
+		if (level.isClientSide) return
+
+		var insertedItem = false
 
 		for (direction in directions) {
 			if (entity.item.isEmpty) continue
@@ -22,15 +26,28 @@ class CollectionPlateBlock : BasePlateBlock() {
 			val itemCap = level.getCapability(Capabilities.ItemHandler.BLOCK, blockPos, direction.opposite) ?: continue
 
 			for (slotIndex in 0 until itemCap.slots) {
+				val previousCount = entity.item.count
 				val remainingStack = itemCap.insertItem(slotIndex, entity.item.copy(), false)
 
 				entity.item = remainingStack
+				if (remainingStack.count < previousCount) insertedItem = true
 
 				if (remainingStack.isEmpty) {
 					entity.discard()
 					break
 				}
 			}
+		}
+
+		if (insertedItem) {
+			level.playSound(
+				null,
+				pos,
+				SoundEvents.ITEM_PICKUP,
+				SoundSource.BLOCKS,
+				0.2f,
+				1f
+			)
 		}
 	}